Rename S3FileHook filename to match class name
