+++ airflow_home/plugins/operators/astronomer_s3_key_sensor.py
@@ -8,7 +8,7 @@ from airflow.utils.decorators import apply_defaults
 from airflow.hooks.base_hook import CONN_ENV_PREFIX
 from urllib import quote_plus
 
-from plugins.hooks.astro_s3_hook import S3FileHook
+from plugins.hooks.s3_file_hook import S3FileHook
 from plugins.operators.utils import build_xcom, is_dir
 
 aws_key = os.getenv('AWS_ACCESS_KEY_ID', '')
